silx.gui.widgets: move OverlayMixIn to protected
