+++ src/silx/gui/widgets/ButtonOverlay.py
@@ -1,6 +1,6 @@
 from silx.gui import qt
 
-from .OverlayMixIn import _OverlayMixIn
+from ._OverlayMixIn import OverlayMixIn as _OverlayMixIn
 
 
 class ButtonOverlay(_OverlayMixIn, qt.QPushButton):
