+++ src/silx/gui/widgets/LabelOverlay.py
@@ -1,6 +1,6 @@
 from silx.gui import qt
 
-from .OverlayMixIn import _OverlayMixIn
+from ._OverlayMixIn import OverlayMixIn as _OverlayMixIn
 
 
 class LabelOverlay(_OverlayMixIn, qt.QLabel):
