+++ src/silx/gui/widgets/WaitingOverlay.py
@@ -1,6 +1,6 @@
 from silx.gui.widgets.WaitingPushButton import WaitingPushButton
 from silx.gui import qt
-from .OverlayMixIn import _OverlayMixIn
+from ._OverlayMixIn import OverlayMixIn as _OverlayMixIn
 
 
 class WaitingOverlay(_OverlayMixIn, WaitingPushButton):
